Torch spmv example working
(this version without sparse tensor decomposition)
